backend/requirements.txt
django==5.1.*
djangorestframework==3.15.*
psycopg2-binary==2.9.*
celery==5.3.*
redis==5.0.*
django-celery-beat>=2.7.0
django-cors-headers==4.3.*
python-decouple==3.8
dj-database-url==2.1.*
pytest-django==4.8.*
factory-boy==3.3.*
